backend/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6
yt-dlp==2024.3.10
ffmpeg-python==0.2.0
openai-whisper==20231117
supabase==2.3.4
python-dotenv==1.0.0
pydantic==2.5.3
httpx==0.26.0
groq==0.4.2
